[mac] Set ViewModel to null on removed
Imagine properties A and B with the same editor type. Both properties
are initially shown and then removed as the selected object is cleared.
The order in which the editors are queued and dequeued for reuse is an
unknown, so we can not depend on a behavior. Imagine then that property
A is re-shown, but not B. It's possible that the editor for B is reused
for property A. Without the change, you now have two editors hooked up
to A's events. Besides keeping the VM alive, if any of the events
trigger UI tasks that require a superview/window, they will fail and
crash.